+++ src/takoyaki/impl/renderer_impl.h
@@ -52,7 +52,7 @@ namespace Takoyaki
         std::unique_ptr<IndexBufferImpl> createIndexBuffer(uint8_t*, EFormat, uint_fast32_t);
         std::unique_ptr<InputLayoutImpl> createInputLayout(const std::string&);
         std::unique_ptr<RootSignatureImpl> createRootSignature(const std::string&);
-        std::unique_ptr<VertexBufferImpl> createVertexBuffer(uint8_t*, uint_fast32_t, uint_fast64_t);
+        std::unique_ptr<VertexBufferImpl> createVertexBuffer(uint8_t*, uint_fast32_t, uint_fast32_t);
 
         void createPipelineState(const std::string&, const PipelineStateDesc&);
